Refactor server to load React
