+++ package.json
@@ -4,9 +4,12 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "start": "node index.js",
+    "start-server": "node server/index.js",
+    "start-react": "react-scripts start",
+    "build-react": "react-scripts build",
     "test": "jest"
   },
+  "proxy": "http://localhost:5001",
   "keywords": [],
   "author": "",
   "license": "ISC",
@@ -17,14 +20,39 @@
     ]
   },
   "dependencies": {
+    "@testing-library/jest-dom": "^4.2.4",
+    "@testing-library/react": "^9.3.2",
+    "@testing-library/user-event": "^7.1.2",
+    "@types/jest": "^24.0.0",
+    "@types/node": "^12.0.0",
+    "@types/react": "^16.9.0",
+    "@types/react-dom": "^16.9.0",
     "axios": "^0.20.0",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
-    "nunjucks": "^3.2.2"
+    "nunjucks": "^3.2.2",
+    "react": "^17.0.0",
+    "react-dom": "^17.0.0",
+    "react-scripts": "3.4.4",
+    "typescript": "~3.7.2"
   },
   "devDependencies": {
     "cross-env": "^7.0.2",
-    "jest": "^26.6.1",
+    "ejs": "^3.1.5",
+    "eslint-plugin-react-hooks": "^4.2.0",
+    "jest": "^24.9.0",
     "supertest": "^5.0.0"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
   }
 }
